# perf(core) Recurrent installs optimization