--- a/packages/yarnpkg-core/sources/index.ts
+++ b/packages/yarnpkg-core/sources/index.ts
@@ -18,6 +18,7 @@ export type {ConfigurationValueMap, ConfigurationDefinitionMap}                 
 export type {Fetcher, FetchOptions, FetchResult, MinimalFetchOptions}               from './Fetcher';
 export {BuildType}                                                                  from './Installer';
 export type {Installer, BuildDirective, InstallStatus, FinalizeInstallStatus}       from './Installer';
+export type {PackageMeta}                                                           from './Installer';
 export {LightReport}                                                                from './LightReport';
 export type {Linker, LinkOptions, MinimalLinkOptions}                               from './Linker';
 export {Manifest}                                                                   from './Manifest';