… .yarn/versions/d848165f.yml
… CHANGELOG.md
… packages/acceptance-tests/pkg-tests-specs/sources/node-modules.test.ts
… packages/yarnpkg-core/sources/Project.ts