**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

This PR speed ups recurrent installs

**How did you fix it?**
<!-- A detailed description of your implementation. -->

Each linker can cache usage of `fetchResult.packageFs` now  (it is an expensive operation) through new interface functions. The core will cache package meta produced by linkers and store it inside `install-state` and will pass it on to the linker on next recurrent install, so that linker can avoid accessing `fetchResult.packageFs`.

| | before | after |
|--|--|--|
| Recurrent Gatsby PnP Linker benchmark | 5.379s | 2.394s |
| Recurrent Gatsby NM Linker benchmark |  5.829s | 3.845s |

**Checklist**
<!--- Don't worry if you miss something, chores are automatically tested. -->
<!--- This checklist exists to help you remember doing the chores when you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).

<!-- See https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check -i` -->
- [x] I have set the packages that need to be released for my changes to be effective.

<!-- The "Testing chores" workflow validates that your PR follows our guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets reviewed.