@larixer 2020-10-24
@crubier @borekb would you be willing to test this PR, I remember you were concerned about install performance. This PR speed ups link step on recurrent installs (`yarn add`) or when you are not installing anything (`yarn`)