@borekb 2020-10-24
Hi @larixer, I sacrificed a lot of my data limit to try this out 😄 but am happy to report that it is indeed faster – repeated `yarn` used to be **~2.7 seconds** on Yarn 2.2.0 and are now **~1.4 s** as of this PR.