@merceyz 2020-10-24
Just to leave it here as well:
Tested it on https://github.com/storybookjs/storybook where they have **4314 dependencies**, there it goes from **~10 seconds** to **~2.5 seconds**