@crubier 2020-10-27
@larixer sorry I missed you initial message.

Tested, on our internal monorepo with 6000 dependencies and 200 packages, it takes the recurrent install from 24s to 13s consistently 🎉 , good job! 

Btw, not the first time I benefit from this, but good job to the team (and @arcanis in particular I think?) on making it super easy to `set version from source` and test easily specific developments!