@merceyz 2020-11-08
Closing since this was included in https://github.com/yarnpkg/berry/pull/2093